# fix(gatsby): reduce page-renderer-size