--- a/packages/gatsby/cache-dir/static-entry.js
+++ b/packages/gatsby/cache-dir/static-entry.js
@@ -6,9 +6,10 @@ const {
   pipeToNodeWritable,
 } = require(`react-dom/server`)
 const { ServerLocation, Router, isRedirect } = require(`@gatsbyjs/reach-router`)
-const { merge, flattenDeep, replace } = require(`lodash`)
+const merge = require(`deepmerge`)
 const { StaticQueryContext } = require(`gatsby`)
 const fs = require(`fs`)
+const { WritableAsPromise } = require(`./server-utils/writable-as-promise`)
 
 const { RouteAnnouncerProps } = require(`./route-announcer-props`)
 const { apiRunner, apiRunnerAsync } = require(`./api-runner-ssr`)
@@ -65,29 +66,40 @@ const getAppDataUrl = () =>
 const createElement = React.createElement
 
 export const sanitizeComponents = components => {
-  const componentsArray = ensureArray(components)
+  const componentsArray = [].concat(components).flat(Infinity)
+
   return componentsArray.map(component => {
     // Ensure manifest is always loaded from content server
     // And not asset server when an assetPrefix is used
     if (__ASSET_PREFIX__ && component.props.rel === `manifest`) {
       return React.cloneElement(component, {
-        href: replace(component.props.href, __ASSET_PREFIX__, ``),
+        href: component.props.href.replace(__ASSET_PREFIX__, ``),
       })
     }
     return component
   })
 }
 
-const ensureArray = components => {
-  if (Array.isArray(components)) {
-    // remove falsy items and flatten
-    return flattenDeep(
-      components.filter(val => (Array.isArray(val) ? val.length > 0 : val))
-    )
-  } else {
-    // we also accept single components, so we need to handle this case as well
-    return components ? [components] : []
+function deepMerge(a, b) {
+  const combineMerge = (target, source, options) => {
+    const destination = target.slice()
+
+    source.forEach((item, index) => {
+      if (typeof destination[index] === `undefined`) {
+        destination[index] = options.cloneUnlessOtherwiseSpecified(
+          item,
+          options
+        )
+      } else if (options.isMergeableObject(item)) {
+        destination[index] = merge(target[index], item, options)
+      } else if (target.indexOf(item) === -1) {
+        destination.push(item)
+      }
+    })
+    return destination
   }
+
+  return merge(a, b, { arrayMerge: combineMerge })
 }
 
 export default async function staticPage({
@@ -149,11 +161,13 @@ export default async function staticPage({
     }
 
     const setHtmlAttributes = attributes => {
-      htmlAttributes = merge(htmlAttributes, attributes)
+      // TODO - we should remove deep merges
+      htmlAttributes = deepMerge(htmlAttributes, attributes)
     }
 
     const setBodyAttributes = attributes => {
-      bodyAttributes = merge(bodyAttributes, attributes)
+      // TODO - we should remove deep merges
+      bodyAttributes = deepMerge(bodyAttributes, attributes)
     }
 
     const setPreBodyComponents = components => {
@@ -169,7 +183,8 @@ export default async function staticPage({
     }
 
     const setBodyProps = props => {
-      bodyProps = merge({}, bodyProps, props)
+      // TODO - we should remove deep merges
+      bodyProps = deepMerge({}, bodyProps, props)
     }
 
     const getHeadComponents = () => headComponents
@@ -266,9 +281,6 @@ export default async function staticPage({
       try {
         // react 18 enabled
         if (pipeToNodeWritable) {
-          const {
-            WritableAsPromise,
-          } = require(`./server-utils/writable-as-promise`)
           const writableStream = new WritableAsPromise()
           const { startWriting } = pipeToNodeWritable(
             bodyComponent,