--- a/packages/gatsby/package.json
+++ b/packages/gatsby/package.json
@@ -51,6 +51,7 @@
     "css.escape": "^1.5.1",
     "date-fns": "^2.14.0",
     "debug": "^3.2.7",
+    "deepmerge": "^4.2.2",
     "del": "^5.1.0",
     "detect-port": "^1.3.0",
     "devcert": "^1.1.3",