--- a/packages/gatsby/src/internal-plugins/bundle-optimisations/gatsby-node.js
+++ b/packages/gatsby/src/internal-plugins/bundle-optimisations/gatsby-node.js
@@ -11,12 +11,24 @@ exports.onCreateWebpackConfig = function onCreateWebpackConfig({
   stage,
   actions,
 }) {
+  const objectAssignStub = path.join(__dirname, `polyfills/object-assign.js`)
+
   if (stage === `build-html` || stage === `develop-html`) {
+    actions.setWebpackConfig({
+      resolve: {
+        alias: {
+          // These files are already polyfilled so these should return in a no-op
+          // Stub Package: object.assign & object-assign
+          "object.assign": objectAssignStub,
+          "object-assign$": objectAssignStub,
+          "@babel/runtime/helpers/extends.js$": objectAssignStub,
+        },
+      },
+    })
     return
   }
 
   const noOp = path.join(__dirname, `polyfills/no-op.js`)
-  const objectAssignStub = path.join(__dirname, `polyfills/object-assign.js`)
   const fetchStub = path.join(__dirname, `polyfills/fetch.js`)
   const whatwgFetchStub = path.join(__dirname, `polyfills/whatwg-fetch.js`)