--- a/packages/gatsby/src/utils/webpack.config.js
+++ b/packages/gatsby/src/utils/webpack.config.js
@@ -302,12 +302,12 @@ module.exports = async (
 
   function getMode() {
     switch (stage) {
-      case `build-javascript`:
-        return `production`
       case `develop`:
+        return `development`
+      case `build-javascript`:
       case `develop-html`:
       case `build-html`:
-        return `development` // So we don't uglify the html bundle
+        return `production`
       default:
         return `production`
     }
@@ -531,6 +531,7 @@ module.exports = async (
   }
 
   const isCssModule = module => module.type === `css/mini-extract`
+
   if (stage === `develop`) {
     config.optimization = {
       splitChunks: {
@@ -564,7 +565,13 @@ module.exports = async (
           },
         },
       },
-      minimizer: [],
+      minimize: false,
+    }
+  }
+
+  if (stage === `develop-html` || stage === `build-html`) {
+    config.optimization = {
+      minimize: false,
     }
   }