<!--
  Have any questions? Check out the contributing docs at https://gatsby.dev/contribute, or
  ask in this Pull Request and a Gatsby maintainer will be happy to help :)
-->

<!--
  Is this a blog post? Check out the docs at https://www.gatsbyjs.com/contributing/blog-contributions/, and please mention if the blog post is pre-approved
  by someone from Gatsby.
-->

## Description

<!-- Write a brief description of the changes introduced by this PR -->
Removes lodash from static-entry and changes webpack config to improve performance:

Tested it on blog-starter
- Drops page-renderer from 827kb => 295kb (decrease of 64.33%)
- Perf cold build 1.8s => 1.3s
- Perf warm build (no change): 1s -> 0.5s